feat(map): expose proposal actions and isolate editor state (#87)
* refactor(farm): isolate editor state and placement validation

* feat(map): expose proposal actions on left click

* fix(farm): isolate proposal creation in event callback
