+++ package.json
@@ -12,7 +12,7 @@
     "dev": "node scripts/dev-local.mjs",
     "build": "npm run verify:changelog && vinext build && node scripts/prepare-built-public.mjs",
     "start": "vinext start",
-    "test": "npm run verify:localization && npm run build && node --test tests/rendered-html.test.mjs tests/localization.test.mjs tests/changelog.test.mjs tests/release-notes.test.mjs tests/mod-compatibility.test.mjs tests/route-planner.test.mjs tests/production-engine.test.mjs tests/forestry-engine.test.mjs tests/machine-engine.test.mjs tests/animal-engine.test.mjs tests/pond-engine.test.mjs tests/portfolio-engine.test.mjs tests/local-package.test.mjs tests/fishing-snapshot.test.mjs tests/mod-consumers.test.mjs tests/mod-production-extractor.test.mjs tests/dashboard-modules.test.mjs tests/dashboard-navigation.test.mjs tests/machine-identity.test.mjs",
+    "test": "npm run verify:localization && npm run build && node --test tests/rendered-html.test.mjs tests/localization.test.mjs tests/changelog.test.mjs tests/release-notes.test.mjs tests/mod-compatibility.test.mjs tests/route-planner.test.mjs tests/production-engine.test.mjs tests/forestry-engine.test.mjs tests/machine-engine.test.mjs tests/animal-engine.test.mjs tests/pond-engine.test.mjs tests/portfolio-engine.test.mjs tests/local-package.test.mjs tests/fishing-snapshot.test.mjs tests/mod-consumers.test.mjs tests/mod-production-extractor.test.mjs tests/dashboard-modules.test.mjs tests/dashboard-navigation.test.mjs tests/farm-editor-interaction.test.mjs tests/machine-identity.test.mjs",
     "verify:changelog": "node scripts/verify-changelog.mjs",
     "verify:localization": "node scripts/verify-localization-ui.mjs",
     "verify:public": "node scripts/verify-public-safety.mjs",
